+++ integrations/github/helpers/utils.py
@@ -1,5 +1,6 @@
 from enum import StrEnum
 
+
 class ObjectKind(StrEnum):
     REPOSITORY = "repository"
     PULL_REQUEST = "pull-request"