+++ integrations/github/webhook_processors/pull_request.py
@@ -1,7 +1,7 @@
 from loguru import logger
-from client import GitHubClient
 from consts import PULL_REQUEST_DELETE_EVENTS, PULL_REQUEST_UPSERT_EVENTS
 from helpers.utils import ObjectKind
+from client import GitHubClient
 from webhook_processors.abstract import GitHubAbstractWebhookProcessor
 from port_ocean.core.handlers.port_app_config.models import ResourceConfig
 from port_ocean.core.handlers.webhook.webhook_event import (
@@ -10,11 +10,13 @@ from port_ocean.core.handlers.webhook.webhook_event import (
     WebhookEventRawResults,
 )
 
+
 class PullRequestWebhookProcessor(GitHubAbstractWebhookProcessor):
     async def should_process_event(self, event: WebhookEvent) -> bool:
+        event_type = event.payload.get("action")
         return (
-            event.payload.get("action") in PULL_REQUEST_UPSERT_EVENTS
-            or event.payload.get("action") in PULL_REQUEST_DELETE_EVENTS
+            event.headers.get("X-GitHub-Event") == "pull_request"
+            and event_type in PULL_REQUEST_UPSERT_EVENTS + PULL_REQUEST_DELETE_EVENTS
         )
 
     async def get_matching_kinds(self, event: WebhookEvent) -> list[str]:
@@ -23,19 +25,28 @@ class PullRequestWebhookProcessor(GitHubAbstractWebhookProcessor):
     async def handle_event(
         self, payload: EventPayload, resource_config: ResourceConfig
     ) -> WebhookEventRawResults:
-        client = GitHubClient.from_ocean_config()
+        action = payload.get("action")
         pr = payload.get("pull_request", {})
-        repo = payload.get("repository", {})
-        
-        if payload.get("action") in PULL_REQUEST_DELETE_EVENTS:
-            logger.info(f"Pull request #{pr.get('number')} was deleted in {repo.get('name')}")
+        repo_name = payload.get("repository", {}).get("name")
+        pr_number = pr.get("number")
+
+        logger.info(
+            f"Processing pull request event: {action} for PR #{pr_number} in {repo_name}"
+        )
+
+        if action in PULL_REQUEST_DELETE_EVENTS:
+            logger.info(f"Pull request #{pr.get('number')} was deleted in {repo_name}")
+
             return WebhookEventRawResults(
                 updated_raw_results=[],
                 deleted_raw_results=[pr],
             )
 
-        logger.info(f"Got event for PR #{pr.get('number')}: {payload.get('action')}")
+        client = GitHubClient.from_ocean_config()
+        latest_pr = await client.get_single_resource(
+            ObjectKind.PULL_REQUEST, f"{repo_name}/{pr_number}"
+        )
+
         return WebhookEventRawResults(
-            updated_raw_results=[pr],
-            deleted_raw_results=[],
-        ) 
+            updated_raw_results=[latest_pr], deleted_raw_results=[]
+        )